.history/requirements_20220127153551.txt
absl-py==1.0.0
adal==1.2.7
aiofiles==0.7.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==2.0.0
aiosignal==1.2.0
altair==4.1.0
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.4.0
appdirs==1.4.4
apturl==0.5.1
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
attrs==21.2.0
azure-common==1.1.25
azure-core==1.7.0
azure-cosmos==4.0.0b6
azure-datalake-store==0.0.52
azure-identity==1.4.0
azure-keyvault-secrets==4.2.0
azure-mgmt-compute==13.0.0
azure-mgmt-core==1.2.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==20.0.0
azure-mgmt-sql==0.20.0
azure-nspkg==3.0.2
azure-storage-blob==12.4.0
Babel==2.9.1
backcall==0.2.0
backoff==1.11.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.8.2
bleach==4.1.0
blessings==1.7
blinker==1.4
boto3==1.20.13
botocore==1.23.13
breezy==3.2.1
Brlapi==0.7.0
cachetools==4.2.4
catkin-pkg==0.4.24
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.8
clearml==1.1.4
cliapp==1.20180812.1
click==7.1.2
cloudpickle==2.0.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==5.1.0
cryptography==36.0.0
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.24
dataclasses-json==0.5.6
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.3
distro==1.6.0
distro-info==1.0
dlib==19.22.1
dnspython==2.1.0
docker-pycreds==0.4.0
docutils==0.18.1
dparse==0.5.1
dulwich==0.20.26
duplicity==0.8.12.0
empy==3.3.4
entrypoints==0.3
et-xmlfile==1.1.0
eventlet==0.33.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==0.1.11
fasteners==0.16.3
fastimport==0.9.14
fiftyone==0.14.2
fiftyone-brain==0.7.2
fiftyone-db==0.3.0
filelock==3.4.0
fin-news==0.1.1
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-Simple-GeoIP==0.2.4
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.2.0
furl==2.1.3
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
glob2==0.7
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-automl==2.5.2
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.53.0
gpg===1.13.1-unknown
gpustat==0.6.0
greenlet==1.1.2
grpcio==1.42.0
gTTS==2.2.3
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
hiredis==2.0.0
html5lib==1.0.1
httpcore==0.14.3
httplib2==0.20.2
httpx==0.21.1
humanfriendly==10.0
identify==2.4.0
idna==2.10
imageio==2.12.0
imageio-ffmpeg==0.4.5
imbalanced-learn==0.8.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
install==1.3.4
iopath==0.1.9
ipaddress==1.0.23
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.12.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.3.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.15.1
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==12.0.0
libcst==0.3.23
libretranslatepy==2.1.1
libvirt-python==6.1.0
livereload==2.6.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.4
macaroonbakery==1.3.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.23.1
monotonic==1.6
motor==2.5.1
moviepy==1.0.3
msal==1.16.0
msal-extensions==0.2.2
msgpack==1.0.3
msrest==0.6.21
msrestazure==0.6.4
multidict==5.2.0
multitasking==0.0.10
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
ndjson==0.3.1
nest-asyncio==1.5.1
netifaces==0.11.0
networkx==2.6.3
nltk==3.6.5
nodeenv==1.6.0
notebook==6.4.6
numpy==1.21.4
nvidia-ml-py==11.450.51
nvidia-ml-py3==7.352.0
nvitop==0.5.1.4
oauth2client==4.1.3
oauthlib==3.1.1
olefile==0.46
opencensus==0.8.0
opencensus-context==0.1.2
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.8.0
parso==0.8.2
pathlib2==2.3.6
pathspec==0.9.0
pathtools==0.1.2
patiencediff==0.2.2
patool==1.12
pexpect==4.8.0
phonenumbers==8.12.37
pickleshare==0.7.5
Pillow==8.4.0
pip-review==1.1.0
pipenv==2021.11.23
platformdirs==2.4.0
playsound==1.3.0
plotly==5.4.0
pluggy==1.0.0
portalocker==1.7.1
pprintpp==0.4.0
pre-commit==2.15.0
proglog==0.1.9
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.22
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
py-spy==0.3.11
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.8.2
pydeck==0.7.1
pydot==1.4.2
PyDrive==1.3.1
PyGithub==1.43.7
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.1
Pympler==0.9
PyNaCl==1.3.0
pyodbc==4.0.30
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyre-check==0.9.8
pyre-extensions==0.0.23
pyRFC3339==1.1
pyrsistent==0.18.0
pyte==0.8.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gitlab==2.0.1
python-taint==0.42
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywatchman==1.4.1
PyWavelets==1.2.0
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
qtconsole==5.2.1
QtPy==1.11.2
ray==1.8.0
redis==4.0.2
regex==2021.11.10
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.5.0
roman==2.0.0
rosdep==0.18.0
rosdistro==0.8.0
rospkg==1.2.3
rsa==4.8
s3transfer==0.5.0
safety==1.10.3
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentry-sdk==1.5.0
setuptools-scm==6.3.2
shortuuid==1.0.8
simple-geoip==0.1.1
simplejson==3.16.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==1.9.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.12
streamlit==1.2.0
stripe==2.63.0
subprocess32==3.5.4
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorrt==0.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thefuck==3.32
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch-tb-profiler==0.3.1
tornado==6.1
tox==3.24.4
tqdm==4.62.3
traitlets==5.1.1
translate==3.6.1
ttystatus==0.38
twilio==6.63.0
typed-ast==1.5.0
typing-inspect==0.7.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
universal-analytics-python3==1.1.1
uritemplate==4.1.1
urllib3==1.25.11
usb-creator==0.3.7
uvicorn==0.11.3
validators==0.18.2
vboxapi==1.0
vcstools==0.1.42
virtualenv==20.10.0
virtualenv-clone==0.5.7
voxel51-eta==0.6.1
wadllib==1.3.3
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==3.5.2
wrapt==1.13.3
wsproto==0.15.0
wstool==0.1.18
xeus-python==0.13.5
xeus-python-shell==0.1.7
xkit==0.0.0
xmltodict==0.12.0
yacs==0.1.8
yarl==1.7.2
yaspin==2.1.0
yfinance==0.1.69
zipp==3.6.0
